[GH-2] Rename folder
